+++ README.md
@@ -35,10 +35,10 @@ mvn clean install
 ```
 
 Try the following examples to learn more about this SDK:
-* [Hello world](./examples/src/test/java/io/mosn/layotto/examples/helloworld)
-* [State management](./examples/src/test/java/io/mosn/layotto/examples/state)
-* [Pubsub API](./examples/src/test/java/io/mosn/layotto/examples/pubsub)
-* [File API](./examples/src/test/java/io/mosn/layotto/examples/file)
+* [Hello world](./examples/src/main/java/io/mosn/layotto/examples/helloworld)
+* [State management](./examples/src/main/java/io/mosn/layotto/examples/state)
+* [Pubsub API](./examples/src/main/java/io/mosn/layotto/examples/pubsub)
+* [File API](./examples/src/main/java/io/mosn/layotto/examples/file)
 
 ## java sdk developer guide
