+++ examples/src/main/java/io/mosn/layotto/examples/pubsub/springboot/Application.java
