+++ examples/src/main/resources/application.properties
